requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
python-libxdo @ https://github.com/rshk/python-libxdo/tarball/master; sys_platform=="linux"
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
urllib3==1.26.2
websockets==8.1
